Move SQ8↔FP16 AVX-512 dispatch to AVX512F tier + flatten F16C guards [MOD-14954]
Two related cleanups in the SQ8↔FP16 dispatch path:

1. The AVX-512 SQ8↔FP16 kernel only uses AVX-512F instructions
   (`_mm512_cvtph_ps`, `_mm512_fmadd_ps`, etc.) but was registered under
   the VNNI tier (`OPT_AVX512_F_BW_VL_VNNI` + check of avx512f/bw/vl/vnni).
   That meant CPUs with AVX-512F but no VNNI (Skylake-X, some Cascade Lake
   variants, etc.) would fall through to AVX2_FMA even though they can
   run the AVX-512 kernel.

   Moved the `Choose_SQ8_FP16_{IP,Cosine,L2}_implementation_AVX512F`
   definitions from `functions/AVX512F_BW_VL_VNNI.cpp` to
   `functions/AVX512F.cpp`, with matching header reshuffle. Dispatch
   sites now gate on `OPT_AVX512F` + `features.avx512f`.

2. F16C is a transversal requirement across the non-AVX-512 SQ8↔FP16
   tiers (SSE4, AVX2, AVX2+FMA) — every one of them widens FP16 query
   lanes via `vcvtph2ps`. Per-tier nested `#ifdef OPT_F16C` was hoisted
   into a single outer block around the three ISA branches in
   `IP_SQ8_FP16_GetDistFunc`, `Cosine_SQ8_FP16_GetDistFunc`, and
   `L2_SQ8_FP16_GetDistFunc`.

Verified: 131 SQ8_FP16 release + 115 ASan + 1166 full test_spaces suite.
