+++ src/VecSim/spaces/functions/AVX2.h
@@ -19,5 +19,4 @@ dist_func_t<float> Choose_SQ8_FP32_L2_implementation_AVX2(size_t dim);
 dist_func_t<float> Choose_BF16_IP_implementation_AVX2(size_t dim);
 dist_func_t<float> Choose_BF16_L2_implementation_AVX2(size_t dim);
 
-
 } // namespace spaces
