+++ src/VecSim/spaces/functions/AVX512F.cpp
@@ -11,10 +11,12 @@
 #include "VecSim/spaces/L2/L2_AVX512F_FP16.h"
 #include "VecSim/spaces/L2/L2_AVX512F_FP32.h"
 #include "VecSim/spaces/L2/L2_AVX512F_FP64.h"
+#include "VecSim/spaces/L2/L2_AVX512F_SQ8_FP16.h"
 
 #include "VecSim/spaces/IP/IP_AVX512F_FP16.h"
 #include "VecSim/spaces/IP/IP_AVX512F_FP32.h"
 #include "VecSim/spaces/IP/IP_AVX512F_FP64.h"
+#include "VecSim/spaces/IP/IP_AVX512F_SQ8_FP16.h"
 
 namespace spaces {
 
@@ -53,6 +55,25 @@ dist_func_t<float> Choose_FP16_IP_implementation_AVX512F(size_t dim) {
 dist_func_t<float> Choose_FP16_L2_implementation_AVX512F(size_t dim) {
     dist_func_t<float> ret_dist_func;
     CHOOSE_IMPLEMENTATION(ret_dist_func, dim, 32, FP16_L2SqrSIMD32_AVX512);
+    return ret_dist_func;
+}
+
+// SQ8↔FP16 kernels only use AVX-512F (cvtph_ps + FMA), so they register here rather than under
+// the VNNI tier — CPUs with AVX-512F but no VNNI (Skylake-X, some Cascade Lake variants) can use
+// these kernels.
+dist_func_t<float> Choose_SQ8_FP16_IP_implementation_AVX512F(size_t dim) {
+    dist_func_t<float> ret_dist_func;
+    CHOOSE_IMPLEMENTATION(ret_dist_func, dim, 16, SQ8_FP16_InnerProductSIMD16_AVX512F);
+    return ret_dist_func;
+}
+dist_func_t<float> Choose_SQ8_FP16_Cosine_implementation_AVX512F(size_t dim) {
+    dist_func_t<float> ret_dist_func;
+    CHOOSE_IMPLEMENTATION(ret_dist_func, dim, 16, SQ8_FP16_CosineSIMD16_AVX512F);
+    return ret_dist_func;
+}
+dist_func_t<float> Choose_SQ8_FP16_L2_implementation_AVX512F(size_t dim) {
+    dist_func_t<float> ret_dist_func;
+    CHOOSE_IMPLEMENTATION(ret_dist_func, dim, 16, SQ8_FP16_L2SqrSIMD16_AVX512F);
     return ret_dist_func;
 }
