+++ src/VecSim/spaces/functions/AVX512F.h
@@ -20,4 +20,9 @@ dist_func_t<float> Choose_FP16_L2_implementation_AVX512F(size_t dim);
 dist_func_t<float> Choose_FP32_L2_implementation_AVX512F(size_t dim);
 dist_func_t<double> Choose_FP64_L2_implementation_AVX512F(size_t dim);
 
+// SQ8↔FP16 kernels — only need AVX-512F, not VNNI/BW/VL.
+dist_func_t<float> Choose_SQ8_FP16_IP_implementation_AVX512F(size_t dim);
+dist_func_t<float> Choose_SQ8_FP16_Cosine_implementation_AVX512F(size_t dim);
+dist_func_t<float> Choose_SQ8_FP16_L2_implementation_AVX512F(size_t dim);
+
 } // namespace spaces
