+++ src/VecSim/spaces/functions/SSE4.h
@@ -16,5 +16,4 @@ dist_func_t<float> Choose_SQ8_FP32_IP_implementation_SSE4(size_t dim);
 dist_func_t<float> Choose_SQ8_FP32_Cosine_implementation_SSE4(size_t dim);
 dist_func_t<float> Choose_SQ8_FP32_L2_implementation_SSE4(size_t dim);
 
-
 } // namespace spaces
